Moved `depth` from `Vertex` into `Node`.
This change shrinks the size of `Vertex`, it does however make the `Descendants` iterator more complex. The iterator now needs to compute the depth of the iterated nodes.

Computing the depth is either done by

1. storing the depth of each `Vertex`. This is fast but means larger memory footprints for `Tree`s.
2. allocating a `Vec` for each iteration to keep track of the current path within the tree.
3. scanning the past `Vertex` entries on each step in the iteration to find the current path in the tree.

I chose to move from option 1 to option 3, meaning we now iterate over previous `Vertex` entries on each step in the iteration. This is probably a slow approach, but entirely avoids allocations. Benchmarking will be needed on various shapes of trees to consider option 2. The current implementation can also easily be optimized to avoid a full scan on every step.