Added cheap `into_sapling()` to turn `Tree`s and `PolyTree`s back into editable saplings.
To avoid allocations the sapling field `path: Vec<usize>` is now also stored in `Tree` and `PolyTree` after the sapling is built. This field is currently never used by the tree types. Turning a tree back into a sapling is then able to reuse the old `path` buffer. This buffer grows to the size of the maximum depth of the sapling.