Reworked internal structure of `Node`.
The changes include removing depth information for nodes; this is a breaking API change. I plan on using a nodes rank instead of its depth. The rank is the nodes position in the trees vertex buffer. Queries should be able to use a nodes rank and scope; the scope being its rank plus its length; instead of relying on depth information. Ranks and lengths are much easier calculate than depth.

The changes to nodes also allows nodes to find their parent node. This is not yet implemented, but now possible. Before this change, a node only had a slice to its subtree, which could not be extended again. Getting the parent would require access to the original tree to get a larger slice. Nodes now always keep a full slice of the trees vertices.

A nodes rank should also make it easy to fetch it from a tree again, meaning we can identify a node using a simple `usize` index. This indexing into the tree is not yet implemented. This commit only contains the change in internal structure.