Added new `Ancestors` iterator.
This double ended iterator takes advantage of the new internal structure of `Node`. It allows getting a nodes parent, or getting the root node of the tree the node is in.

Both directions of the iterator are linear speed. The iterator has to chew through a lot of vertices to get to the next relevant ancestor. It does however natively offer searching backwards or forwards, making it double ended. It can quickly find the parent of a node, or the root of the tree.

Using the iterator to count all ancestor is a slow method of determining the depth of a node. This actually is how `Descendants` used to determine its yielded nodes depth. There currently is no fast alternative to computing a nodes depth.