Added `Node::parent`.
The new ancestors iterator made it easy to get the parent node of any node. The iterator searches for the first vertex containing the node backwards from the nodes vertex. I cannot think of a faster implementation. The runtime depends only on the number of vertices between the node and its parent; this is the difference between their ranks. If the node is one of the later child nodes of its parent, especially if the earlier children have large subtrees, this difference in rank may be significant.