Added `Node::prune`.
This is a replacement for the planned `SubTree` type. When pruning a node a new node is returned that is no longer aware of the entire original tree. The pruned node only knows about itself. This property is inherited when iterating over nodes. All descendants of the pruned node will only be aware of the pruned node, not the entire tree. This was planned to be represented by its own type `SubTree`, but I decided the type would be too similar to `Node`.

This change also requires `Node::get` to get nodes by rank; this method was only available on `Tree`s before. The rank may not be relative to a trees root anymore, but instead to a pruned node. The user needs to use `get` on the pruned node or any of its descendants. Using a rank on the wrong node or tree will return some unexpected other node.