Changed `Sapling::pop` to return `&mut T` instead of `&T`.
This is technically API breaking, however I cannot think of any code using the crate that would change its behavior or break; so maybe it is not.

Also added `Sapling::peek_mut` that acts similar to `Sapling::peek` but returns `&mut T` and requires `&mut self`.

This change was motivated by the need to update a nodes payload after it was added to the sapling. The payload can now be mutated or replaced when the node is closed.