Added temporary build values to `Sapling`.
The sapling struct now has two generics, `Sapling<T, TMP = ()>`. The new generic defaults to `()` to ensure user code will continue to function the same. This change may actually not be API breaking; at least my tests still work.

There are at least two API breaking changes. The methods `Sapling::push_tree` and `Sapling::push_polytree` are now generic. I think all other changes are non breaking changes.

It is now possible to construct a sapling with an arbitrary type for temporary values, which are stored for open nodes during construction of a tree. These values help in computing a nodes final payload when closing it. When pushing a new node onto the sapling a temporary value is set. This value can be mutated by peeking the node. When popping the node the temporary value is returned and allows computing and setting the nodes payload through a `&mut T` using `std::mem::replace`.

This functionality may be of interest for my `read-xml` crate, though I no longer plan on pursuing the idea that required this change.

This new functionality may be better suited to a separate type of sapling, such as `Sapling2<T, TMP>`. I was curious to see if the type system allows to squeeze this functionality into the existing `Sapling<T>` without bothering users that are not interested in using temporary values.

The temporary values are stored in a new `path_data: Vec<TMP>`. This vec is in sync with the old field `path: Vec<usize>`. I kept these buffers separate so that `path` is moved into the `Tree<T>` that is built from the sapling, whereas `path_data` is dropped and recreated when the tree is turned back into a sapling. Assuming that `TMP = ()` the recreating of `path_data` will not cause any allocations because `()` is zero sized. It may make more sense to make trees generic for `TMP` as well, at which point the two path fields should probably be merged into `path: Vec<(usize, TMP)>`. I mostly want to avoid nodes and iterators gaining the second generic.